pyproject.toml
[tool.poetry]
name = "v1-service-commit"
version = "0.1.0"
description = ""
authors = ["hwc9169 <hwc9169@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.72.0"
aioredis = "^2.0.1"
uvicorn = "^0.17.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
